LayoutTests/storage/filesystemaccess/sync-access-handle-storage-limit-worker.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
if (window.testRunner)
    testRunner.setAllowStorageQuotaIncrease(false);
worker = startWorker('resources/sync-access-handle-storage-limit.js');</script>
</body>
</html>
